requirements.txt
# Core ML and AI libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Vector database and search
faiss-cpu>=1.7.4

# Document processing
pypdf2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
requests>=2.31.0

# Web interface
streamlit>=1.25.0
plotly>=5.15.0

# Export functionality
reportlab>=4.0.0
markdown>=3.4.0

# CLI enhancements
rich>=13.0.0

# Utilities
tqdm>=4.65.0
pathlib2>=2.3.7

# Optional: GPU support (uncomment if you have CUDA)
# faiss-gpu>=1.7.4
# torch-gpu>=2.0.0
